Fix option parsing for intentionally-blank string values
Properly handle string options with required values if a blank string is
supplied. Previously these did not work, but now the following formats are
handled properly:
    --long-opt=''
    --long-opt ''
    -l ''

Note that -l'' (without a space) is not supported, as the shell makes this
impossible anyway.

The practical usage for supplying a blank-string is to allow overriding a non-
blank option value set by a lower-priority option source.