Better handling of skip- prefix, and differentiate vs empty strings
NormalizeOptionToken now returns an extra value, allowing the caller to
distinguish options that have no value specified at all, vs ones that are
explicitly set to an empty string or used the skip- prefix. (Potentially
breaking change, since NormalizeOptionToken is exported.)

When skip- prefix is used for a string option, interpret the value as ""
instead of as "0". (Ditto for bool options, but Config.GetBool handles "" and
"0" equivalently, so no effective change.)

Change how usage text displays boolean options. Ones that default to false now
display as "--option-name", and ones that default to true now display as
"--[skip-]option-name", to make intended usage clearer.

Disallow use of Option.ValueRequired() on bool options, since it does not make
sense to permit and other code makes assumptions about this.