ParseCLI: panic upon duplicate option shorthands
Previously, if two Options in the same Command had the same Shorthand rune,
the behavior was undefined: the short-form (single-dash) option parsing for
that rune on the CLI would randomly correspond to whichever conflicting Option
was processed last, based on Golang's random map iteration order.

This commit adds a check to ParseCLI() to explicitly panic in case of
conflict, since this situation is indicative of programmer error.