new method Config.GetAllowEnvVar
This commit adds a new Config method, GetAllowEnvVar, which supports
environment variable lookups: if the value retrieved from the configuration
looks like $FOO or "$FOO", the returned value will be that of the
corresponding ENV var value, even if it is blank / not set.

Environment value lookups only occur if the option value came from a source
other than the CLI, since it is assumed that shells already handle the CLI
use-case appropriately. The value must also either not be quote-wrapped, or
be wrapped in double-quotes. (This way, literal values that just happen to
begin with a dollar sign may be expressed by wrapping a string in single-
quotes.)