new method Config.GetAbsPath
This commit adds a new Config method for returning an option value interpreted
as a filesystem path. If the option value is blank or already an absolute
path, it is returned as-is. If the option value is a relative path and it was
set in an option file, it is interpreted relative to that option file's path.
Otherwise, relative paths set in other locations (command-line, option default
value, runtime override) are interpreted relative to the working directory.