file: add support for limiting which options are processed
This commit adds the ability to limit which options are processed in an
option file. When used, only the specified option names will be processed,
and all other options will be ignored. This is useful when integrating with
option files from other software, where only some specific options are
relevant for processing.